pyproject.toml
[tool.poetry]
name = "mastermind-twitch"
version = "0.1.0"
description = "Mastermind game for twitch chat"
authors = ["Walfask"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
twitchio = "^2.6.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.280"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["F", "E", "W", "I001"]
line-length = 120
fixable = ["ALL"]

src = ["src"]
